test/requirement.txt
alembic==0.9.6
blinker==1.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
pymongo==3.5.1
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
pywin32==221
six==1.11.0
SQLAlchemy==1.1.14
tornado==4.5.2
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
